RS-39 Reorganize CLI and scanner integration tests into modular suites
- CLI and scanner integration tests have been refactored and split into focused modules for improved maintainability and readability.
- Large monolithic test files were replaced with smaller, purpose-driven modules grouped by functionality (e.g., argument parsing, validation, repository handling, commit processing).
- Common test helpers for scanner unit tests were extracted into a dedicated helpers module to reduce duplication.
- Test imports and helper functions were updated to use the new modular structure.
- Documentation comments and module re-exports were added to clarify test organization and usage.
- No changes were made to core application or library logic; all changes are limited to test code structure and organization.
