+++ src/scanner/task/tests/author_filtering.rs
@@ -5,44 +5,11 @@
 // Removed unused import: super::helpers::*
 use super::super::*;
 use crate::core::query::QueryParams;
-use crate::scanner::error::ScanResult;
+use crate::scanner::tests::helpers::{collect_scan_messages, count_commit_messages};
 use crate::scanner::types::{ScanMessage, ScanRequires};
 use serial_test::serial;
 use std::process::{Command, Output};
 use tempfile::TempDir;
-
-/// Test helper to collect scan messages into a Vec using the streaming callback API
-async fn collect_scan_messages(
-    scanner_task: &ScannerTask,
-    query_params: Option<&QueryParams>,
-) -> ScanResult<Vec<ScanMessage>> {
-    use std::cell::RefCell;
-    use std::rc::Rc;
-
-    let messages = Rc::new(RefCell::new(Vec::new()));
-    let messages_clone = Rc::clone(&messages);
-
-    scanner_task
-        .scan_commits_with_query(query_params, move |msg| {
-            messages_clone.borrow_mut().push(msg);
-            async { Ok(()) }
-        })
-        .await?;
-
-    Ok(Rc::try_unwrap(messages).unwrap().into_inner())
-}
-
-/// Test helper to count commit data messages without collecting all messages
-async fn count_commit_messages(
-    scanner_task: &ScannerTask,
-    query_params: Option<&QueryParams>,
-) -> ScanResult<usize> {
-    let messages = collect_scan_messages(scanner_task, query_params).await?;
-    Ok(messages
-        .iter()
-        .filter(|m| matches!(m, ScanMessage::CommitData { .. }))
-        .count())
-}
 
 /// Helper function to run a Git command and ensure it succeeds
 fn run_git_command(args: &[&str], dir: &std::path::Path, error_msg: &str) -> Output {
