+++ tests/cli/mod.rs
@@ -1,0 +1,12 @@
+//! CLI Integration Test Modules
+//!
+//! Organized CLI integration tests split into focused modules for better maintainability.
+
+pub mod argument_parsing;
+pub mod checkout;
+pub mod filtering;
+pub mod path_handling;
+pub mod plugin_config;
+pub mod repository_handling;
+pub mod toml_config;
+pub mod validation;
