+++ tests/scanner/mod.rs
@@ -1,0 +1,9 @@
+//! Scanner Integration Test Modules
+//!
+//! Organized scanner integration tests split into focused modules for better maintainability.
+
+pub mod manager;
+pub mod tasks;
+pub mod repository;
+pub mod messaging;
+pub mod commit_processing;
